bites the bullet and move partials into separate file
