+++ project/npda/views/__init__.py
@@ -4,4 +4,7 @@ from .npda_users import *
 from .patient import *
 from .submissions import *
 from .visit import *
-from .dashboard import *
+# Dashboard view
+from .dashboard.dashboard import dashboard
+# Expose partials here
+from project.npda.views.dashboard.partials import *
